repos/edge_agent/requirements.txt
numpy>=1.26
opencv-python>=4.8
requests>=2.31
python-dotenv>=1.0
